user can add goals

+++ back_end/db/migrate/20191126053641_create_categories.rb
@@ -2,7 +2,7 @@ class CreateCategories < ActiveRecord::Migration[5.2]
   def change
     create_table :categories do |t|
       t.string :name
-      t.integer :cost
+      t.float :cost
       t.references :user
 
       t.timestamps
